backend/requirements.txt
pymilvus
pymilvus[milvus_lite]
pydantic
fastapi
fastapi[standard]
ffmpeg-python
elevenlabs
python-dotenv
google-genai